Rename client/src/panels/intro_graphs to intro_panels, other references to panels as 'graphs' within
