+++ client/src/infographic/get_panels_for_subject/get_crso_panels.js
@@ -2,33 +2,33 @@ import { ensure_loaded } from '../../core/lazy_loader.js';
 
 // shared all
 import { declare_welcome_mat_panel } from '../../panels/welcome_mat/welcome_mat.js';
-import { declare_financial_intro_panel } from '../../panels/intro_graphs/index.js';
+import { declare_financial_intro_panel } from '../../panels/intro_panels/index.js';
 
 // shared gov, dept, crso, program
 import {
   declare_results_intro_panel,
   declare_late_dps_warning_panel,
-} from '../../panels/intro_graphs/index.js';
+} from '../../panels/intro_panels/index.js';
 import { declare_budget_measures_panel } from '../../panels/budget_measures/budget_measures_panel.js';
-import { declare_year_warning_panel } from '../../panels/intro_graphs/index.js';
+import { declare_year_warning_panel } from '../../panels/intro_panels/index.js';
 
 // shared dept, crso, program
 import { declare_explore_results_panel } from '../../panels/result_graphs/index.js';
 import { declare_results_table_panel } from '../../panels/result_graphs/index.js';
-import { declare_tags_of_interest_panel } from '../../panels/intro_graphs/index.js';
+import { declare_tags_of_interest_panel } from '../../panels/intro_panels/index.js';
 import { declare_drr_planned_actual_panel } from "../../panels/drr_dp_resources/drr_planned_actual.js";
 import { declare_dp_rev_split_panel } from "../../panels/drr_dp_resources/dp_rev_split.js";
 import { declare_drr_summary_panel } from '../../panels/result_graphs/index.js';
 
 // shared crso, program
-import { declare_profile_panel } from '../../panels/intro_graphs/index.js';
+import { declare_profile_panel } from '../../panels/intro_panels/index.js';
 
 // crso only panels
 import {
   declare_dead_crso_warning_panel,
   declare_crso_in_gov_panel,
   declare_crso_links_to_other_crso_panel,
-} from '../../panels/intro_graphs/index.js';
+} from '../../panels/intro_panels/index.js';
 import {
   declare_crso_by_prog_fte_panel,
   declare_crso_by_prog_exp_panel,
