+++ client/src/infographic/get_panels_for_subject/get_dept_panels.js
@@ -3,20 +3,20 @@ import { ensure_loaded } from '../../core/lazy_loader.js';
 
 // shared all
 import { declare_welcome_mat_panel } from '../../panels/welcome_mat/welcome_mat.js';
-import { declare_financial_intro_panel } from '../../panels/intro_graphs/index.js';
+import { declare_financial_intro_panel } from '../../panels/intro_panels/index.js';
 
 // shared gov, dept, crso, program
 import {
   declare_results_intro_panel,
   declare_late_dps_warning_panel,
-} from '../../panels/intro_graphs/index.js';
+} from '../../panels/intro_panels/index.js';
 import { declare_budget_measures_panel } from '../../panels/budget_measures/budget_measures_panel.js';
-import { declare_year_warning_panel } from '../../panels/intro_graphs/index.js';
+import { declare_year_warning_panel } from '../../panels/intro_panels/index.js';
 
 // shared dept, crso, program
 import { declare_explore_results_panel } from '../../panels/result_graphs/index.js';
 import { declare_results_table_panel } from '../../panels/result_graphs/index.js';
-import { declare_tags_of_interest_panel } from '../../panels/intro_graphs/index.js';
+import { declare_tags_of_interest_panel } from '../../panels/intro_panels/index.js';
 import { declare_drr_planned_actual_panel } from "../../panels/drr_dp_resources/drr_planned_actual.js";
 import { declare_dp_rev_split_panel } from "../../panels/drr_dp_resources/dp_rev_split.js";
 import { declare_drr_summary_panel } from '../../panels/result_graphs/index.js';
@@ -25,7 +25,7 @@ import { declare_drr_summary_panel } from '../../panels/result_graphs/index.js';
 import { declare_spend_rev_split_panel } from "../../panels/sobj/spend_rev_split.js";
 
 // shared gov, dept
-import { declare_links_to_rpb_panel } from '../../panels/intro_graphs/index.js';
+import { declare_links_to_rpb_panel } from '../../panels/intro_panels/index.js';
 import { declare_in_year_voted_stat_split_panel } from "../../panels/vote_stat/in_year_vote_stat_split.js";
 //import { declare_in_year_estimates_split_panel } from "../../panels/vote_stat/in_year_estimates_split.js";
 import { declare_historical_g_and_c_panel } from "../../panels/transfer_payments/historical_g_and_c.js";
@@ -38,7 +38,7 @@ import { declare_detailed_program_spending_split_panel } from "../../panels/deta
 import {
   declare_portfolio_structure_intro_panel,
   declare_portfolio_structure_related_panel,
-} from '../../panels/intro_graphs/index.js';
+} from '../../panels/intro_panels/index.js';
 import { declare_estimates_in_perspective_panel } from "../../panels/vote_stat/estimates_in_perspective.js";
 import { declare_spend_by_so_hist_panel } from "../../panels/sobj/spend_by_so_hist.js";
 import { declare_igoc_fields_panel } from '../../panels/igoc/igoc_panel.js';
