+++ client/src/infographic/get_panels_for_subject/get_program_panels.js
@@ -2,20 +2,20 @@ import { ensure_loaded } from '../../core/lazy_loader.js';
 
 // shared all
 import { declare_welcome_mat_panel } from '../../panels/welcome_mat/welcome_mat.js';
-import { declare_financial_intro_panel } from '../../panels/intro_graphs/index.js';
+import { declare_financial_intro_panel } from '../../panels/intro_panels/index.js';
 
 // shared gov, dept, crso, program
 import {
   declare_results_intro_panel,
   declare_late_dps_warning_panel,
-} from '../../panels/intro_graphs/index.js';
+} from '../../panels/intro_panels/index.js';
 import { declare_budget_measures_panel } from '../../panels/budget_measures/budget_measures_panel.js';
-import { declare_year_warning_panel } from '../../panels/intro_graphs/index.js';
+import { declare_year_warning_panel } from '../../panels/intro_panels/index.js';
 
 // shared dept, crso, program
 import { declare_explore_results_panel } from '../../panels/result_graphs/index.js';
 import { declare_results_table_panel } from '../../panels/result_graphs/index.js';
-import { declare_tags_of_interest_panel } from '../../panels/intro_graphs/index.js';
+import { declare_tags_of_interest_panel } from '../../panels/intro_panels/index.js';
 import { declare_drr_planned_actual_panel } from "../../panels/drr_dp_resources/drr_planned_actual.js";
 import { declare_dp_rev_split_panel } from "../../panels/drr_dp_resources/dp_rev_split.js";
 import { declare_drr_summary_panel } from '../../panels/result_graphs/index.js';
@@ -24,7 +24,7 @@ import { declare_drr_summary_panel } from '../../panels/result_graphs/index.js';
 import { declare_spend_rev_split_panel } from "../../panels/sobj/spend_rev_split.js";
 
 // shared crso, program
-import { declare_profile_panel } from '../../panels/intro_graphs/index.js';
+import { declare_profile_panel } from '../../panels/intro_panels/index.js';
 
 // shared program, tag
 import { declare_vote_stat_split_panel } from "../../panels/vote_stat/last_year_vote_stat_split.js";
@@ -35,7 +35,7 @@ import {
   declare_dead_program_warning_panel,
   declare_program_fed_structure_panel,
   declare_related_program_structure_panel,
-} from '../../panels/intro_graphs/index.js';
+} from '../../panels/intro_panels/index.js';
 import { declare_spending_in_tag_perspective_panel } from "../../panels/drr_dp_resources/spending_in_perspective.js";
 import { declare_sub_program_resources_panel } from "../../panels/drr_dp_resources/sub_program_resources.js";
