+++ client/src/panels/intro_panels/financial_intro.yaml
