+++ client/src/panels/intro_panels/intro_panel_text_provider.js
