+++ client/src/panels/intro_panels/people_intro.yaml
