+++ client/src/panels/intro_panels/simplographic.js
