+++ client/src/panels/intro_panels/simplographic.yaml
@@ -1,0 +1,56 @@
+simplographic_spending_title:
+  en: How much we spend
+  fr: Combien nous dépensons
+
+simplographic_spending_text:
+  transform: [handlebars,markdown]
+  en: |
+    In **{{pa_last_year}}**, the government spent **<a href="{{org_vote_stat_pa_link}}">{{fmt_compact1_written gov_exp_pa_last_year}}</a>**. The three largest items in terms of spending were:
+
+    {{#each largest_items}} 
+      * {{this.desc}} - {{infograph_link this.subject}}
+    {{/each}}
+  fr : |
+      En **{{pa_last_year}}**, le gouvernement a dépensé **<a href="{{org_vote_stat_pa_link}}">{{fmt_compact1_written gov_exp_pa_last_year}}</a>**. Les trois items suivant représentaient la plus grande partie des dépenses :
+
+      {{#each largest_items}} 
+        * {{this.desc}} - {{infograph_link this.subject}}
+      {{/each}}
+
+
+simplographic_people_title:
+  en: Who we employ
+  fr: Qui employons-nous 
+
+simplographic_people_text:
+  transform: [handlebars,markdown]
+  en: |
+    As of **{{ppl_last_year_end_ticks}}** there were **<a href="{{org_employee_type_link}}">{{fmt_big_int empl_count_total}}</a>** employees in the {{gl_tt "Federal Public Service" "FPS"}}, with **<a href="{{org_employee_region_link}}">{{fmt_percentage1 empl_count_ncr_ratio}}</a>** located in the National Capital Region.
+  fr: |
+    Au **31 mars {{ppl_last_year_short_second}}**, il y avait **<a href="{{org_employee_type_link}}">{{fmt_big_int empl_count_total}}</a>** employés au sein de la {{gl_tt "fonction publique générale" "FPS"}}, dont **<a href="{{org_employee_region_link}}">{{fmt_percentage1 empl_count_ncr_ratio}}</a>** travaillaient dans la région de la capitale nationale.
+
+simplographic_struct_title:
+  en: How we're structured
+  fr: Comment nous sommes structurés
+
+simplographic_struct_text:
+  transform: [handlebars,markdown]
+  en: |
+    The government is made up of **{{num_federal_inst}}** [federal organizations](#igoc/inst_form_filter)
+    across **{{num_ministries}}** portfolios (areas of responsibility).
+  fr: |
+    Le gouvernement est composé de **{{num_federal_inst}}** [organisations fédérales](#igoc/inst_form_filter) et de **{{num_ministries}}** différents portefeuilles (secteurs de responsabilité).
+
+simplographic_results_title:
+  en: What we achieved
+  fr: Ce que nous avons réalisé
+
+simplographic_results_text:
+  transform: [handlebars,markdown]
+  en: |
+    In {{pa_last_year}}, the federal government sought to achieve **{{num_results}}** results. Progress towards meeting these results was measured using **{{fmt_big_int num_indicators}}** {{gl_tt "indicators" "IND_RESULT"}}. 
+    Of those indicators, **{{fmt_big_int num_met}}** (**{{fmt_percentage pct_met}}**) have {{gl_tt "met" "RESULTS_MET"}} their target.
+  fr: |
+    En {{pa_last_year}}, le gouvernement fédéral a cherché à rencontrer **{{num_results}}** résultats. Les progrès accomplis dans l'atteinte de ces résultats ont été mesurés à l’aide de **{{fmt_big_int num_indicators}}** {{gl_tt "indicateurs" "IND_RESULT"}}. 
+    De ces indicateurs, **{{fmt_big_int num_met}}** (**{{fmt_percentage pct_met}}**) ont {{gl_tt "atteint" "RESULTS_MET"}} leur cible.
+
