+++ client/src/panels/intro_panels/tagging_intro.yaml
