+++ client/src/panels/intro_panels/tags_related_to_subject_panels.js
@@ -1,6 +1,6 @@
 import classNames from 'classnames';
 import { Fragment } from 'react';
-import { text_maker, TM } from './intro_graph_text_provider.js';
+import { text_maker, TM } from './intro_panel_text_provider.js';
 import { 
   Subject, 
   declare_panel,
